pyproject.toml
[tool.poetry]
name = "credit_service"
version = "0.1.0"
description = ""
authors = ["Artsiom Zhvanko <azhvanko@outlook.com>"]
readme = "README.md"
packages = [
    {include = "credit_service"}
]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
djangorestframework = "3.14.0"
psycopg2-binary = "2.9.6"
python-environ = "0.4.54"
structlog = "23.1.0"
drf-yasg = "1.21.5"
django-filter = "23.2"

[tool.poetry.group.dev.dependencies]
django-stubs = "4.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
